requirements.txt
# Core web framework
Flask==2.3.3
gunicorn==21.2.0
Werkzeug==2.3.7

# Data science core
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2

# XGBoost
xgboost==1.7.6

# PyTorch (CPU version - required for your disease prediction model)
torch==2.0.1
torchvision==0.15.2

# Image processing
Pillow==10.0.0

# Utilities
requests==2.31.0
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
Jinja2==3.1.2